Resolve #18 Avoid commas as decimal separator
e.g. in German language, php prints commas as decimal separators instead of periods.
This results in broken inline CSS code, defining the height of the image.

I assume the `str_replace()` is probably not the best way to resolve that, but will at least solve the problem for now.